test: move vat-status.js
