+++ packages/deployment/upgrade-test/upgrade-test-scripts/agoric-upgrade-12/post.test.js
@@ -1,5 +1,5 @@
 import test from 'ava';
-import { getIncarnation } from './tools/vat-status.js';
+import { getIncarnation } from '../lib/vat-status.js';
 
 test(`verify Zoe vat incarnation`, async t => {
   const incarantion = await getIncarnation('zoe');
